+++ 02-JPA/03-One2Many-Uni-Fetch-Eager/src/main/java/com/jpa/one2many/uni/eager/dao/UserDaoImpl.java
@@ -9,7 +9,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.jpa.one2many.uni.eager.entity.RoleEntity;
 import com.jpa.one2many.uni.eager.entity.UserEntity;
-import com.jpa.one2many.uni.eager.repository.RoleRepository;
+import com.jpa.one2many.uni.eager.exception.ResourceNotFoundException;
 import com.jpa.one2many.uni.eager.repository.UserRepository;
 
 @Service
@@ -18,9 +18,6 @@ public class UserDaoImpl implements UserDao {
 	@Autowired
 	private UserRepository userRepository;
 
-	@Autowired
-	private RoleRepository roleRepository;
-
 	@Override
 	public UserEntity createUser(UserEntity userEntity) {
 		return userRepository.save(userEntity);
@@ -28,25 +25,38 @@ public class UserDaoImpl implements UserDao {
 
 	@Override
 	public UserEntity getUserById(long id) {
-//		return userRepository.findById(id);
-		return userRepository.jpqlFindById(id);
-//		return userRepository.nativeFindById(id);
+		//	UserEntity userEntity = userRepository.findById(id);
+		//	UserEntity userEntity =  userRepository.nativeFindById(id);
+		UserEntity userEntity = userRepository.jpqlFindById(id);		
+		if(userEntity == null)
+			throw new ResourceNotFoundException("Not found User with id = " + id);			
+		return userEntity;
 	}
 	
 	@Override
-	public UserEntity getUserByPid(long pid) {
-		return userRepository.findByPid(pid);
+	public UserEntity getUserByPid(long pid) {		
+		UserEntity userEntity = userRepository.findByPid(pid);		
+		if(userEntity == null)
+			throw new ResourceNotFoundException("Not found User with pid = " + pid);	
+		return userEntity;
 	}
 	
 	@Override
-	public UserEntity getUserByName(String name) {
-		return userRepository.findByName(name);
+	public UserEntity getUserByName(String name) {			
+		UserEntity userEntity = userRepository.findByName(name);		
+		if(userEntity == null)
+			throw new ResourceNotFoundException("Not found User with name = " + name);	
+		return userEntity;
 	}	
 
 	@Override
 	public UserEntity getUserByEmail(String email) {
-		return userRepository.findByEmail(email);
+		UserEntity userEntity = userRepository.findByEmail(email);		
+		if(userEntity == null)
+			throw new ResourceNotFoundException("Not found User with email = " + email);	
+		return userEntity;
 	}	
+
 
 	@Override
 	public List<UserEntity> getAllUsers() {
@@ -56,20 +66,19 @@ public class UserDaoImpl implements UserDao {
 	@Override
 	public void removeUserByPid(long pid) {
 		UserEntity userEntity = userRepository.findByPid(pid);
+		if(userEntity == null)
+			throw new ResourceNotFoundException("Not found User with pid = " + pid);	
 		userRepository.delete(userEntity);
 	}
 	
 	@Override
 	public UserEntity addRoleToUser(long userPid, RoleEntity roleEntity) {
-
 		UserEntity userEntity = userRepository.findByPid(userPid);
-
+		if(userEntity == null)
+			throw new ResourceNotFoundException("Not found User with userPid = " + userPid);
 		roleEntity.setPid(userPid);
-
 		userEntity.addRole(roleEntity);
-
 		UserEntity returnedValue = userRepository.save(userEntity);
-
 		return returnedValue;
 	}
 
@@ -96,11 +105,12 @@ public class UserDaoImpl implements UserDao {
 
 		UserEntity userEntity = userRepository.findByPid(userPid);
 		
+		if(userEntity == null)
+			throw new ResourceNotFoundException("Not found User with userPid = " + userPid);	
+
 		/**
 		 * there are 4 different ways to retrieve roleEntity from DB
 		 */
-		
-		long start = System.nanoTime();
 		
 		/**
 		 * (1) We Don't need to add @Transactional If we search with For loop  
@@ -131,13 +141,7 @@ public class UserDaoImpl implements UserDao {
 //		RoleEntity roleEntity = roleRepository.findByPidAndRole(userPid, role);
 																											
 		userEntity.removeRole(roleEntity);
-		
-		long end = System.nanoTime();
-		
-		System.out.println(end - start);
-		
 		UserEntity returnedValue = userRepository.save(userEntity);
-		
 		return returnedValue;
 	}
 }
